add qiniu

+++ parsedata/QiniuHttpClient.m
@@ -75,7 +75,7 @@
         complete(operation,error);
     }];
     
-    [operation setUploadProgressBlock:^(NSUInteger bytesWritten, long long totalBytesWritten, long long totalBytesExpectedToWrite) {
+    [operation setUploadProgressBlock:^(NSUInteger bytesWritten, NSInteger totalBytesWritten, NSInteger totalBytesExpectedToWrite) {
         progressBlock((float)totalBytesWritten / (float)totalBytesExpectedToWrite);
     }];
     [self.operationQueue addOperation:operation];
@@ -117,7 +117,7 @@
                                                                           complete(operation,error);
                                                                       }];
     
-    [operation setUploadProgressBlock:^(NSUInteger bytesWritten, long long totalBytesWritten, long long totalBytesExpectedToWrite) {
+    [operation setUploadProgressBlock:^(NSUInteger bytesWritten, NSInteger totalBytesWritten, NSInteger totalBytesExpectedToWrite) {
         progressBlock((float)totalBytesWritten / (float)totalBytesExpectedToWrite);
     }];
     [self.operationQueue addOperation:operation];
